requirements.txt
Django==4.1.3
colorama==0.4.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.3.0
requests==2.28.1
xmltodict==0.13.0
